+++ test/src/editor_test.dart
@@ -12,7 +12,8 @@ Editor editor;
 @Setup
 void setup() {
   html.document.body.append(new html.Element.div()..id = 'editor');
-  editor = edit(html.query('#editor'))..setValue(sampleText, -1);
+  editor = edit(html.query('#editor'))
+      ..setValue(sampleText, cursorPosition: -1);
 }
 
 @Teardown
@@ -83,17 +84,17 @@ void testFocus() {
 void testValue() {
   expect(editor.value, equals(sampleText));
   // 0 = select all
-  editor.setValue('snarf', 0);
+  editor.setValue('snarf');
   expect(editor.value, equals('snarf'));
   expect(editor.cursorPosition, equals(const Point(0,5))); 
   expect(editor.selectionRange, equals(new Range(0,0,0,5)));
   // -1 = document start
-  editor.setValue('start', -1);
+  editor.setValue('start', cursorPosition: -1);
   expect(editor.value, equals('start'));
   expect(editor.cursorPosition, equals(const Point(0,0)));
   expect(editor.selectionRange, equals(new Range(0,0,0,0)));
   // 1 = document end
-  editor.setValue('end', 1);
+  editor.setValue('end', cursorPosition: 1);
   expect(editor.value, equals('end'));
   expect(editor.cursorPosition, equals(const Point(0,3)));
   expect(editor.selectionRange, equals(new Range(0,3,0,3)));
@@ -262,7 +263,7 @@ void testRemoveToLineStart() {
 
 @Test()
 void testRemoveWordLeft() {
-  editor.setValue(sampleText, 1);
+  editor.setValue(sampleText, cursorPosition: 1);
   expect(editor.cursorPosition, equals(const Point(5, 76)));  
   editor.onChange.listen(expectAsync1((Delta delta) {
     expect(delta, const isInstanceOf<RemoveTextDelta>());