+++ CHANGELOG.md
@@ -8,6 +8,9 @@ coverage.
 parameter `clip` which defaults to `true`.
 - Changed the `dontSelect` parameter of `UndoManager.undo` and `redo` to a named 
 optional parameter `select` which defaults to `true`.
+- Changed the `cursorPos` parameter of `Editor.setValue` to a named optional
+parameter `cursorPosition` which defaults to `0` (select all) and documented the
+method.
 
 ## 0.0.6+10.7.2013